+++ torch_points_kernels/__init__.py
@@ -1,6 +1,7 @@
 from .torchpoints import *
 from .knn import knn
 from .cluster import region_grow
+from .metrics import instance_iou
 
 __all__ = [
     "ball_query",
@@ -10,4 +11,5 @@ __all__ = [
     "three_nn",
     "knn",
     "region_grow",
+    "instance_iou",
 ]